Fixed bug #60206 (possible integer overflow in content_length)
git-svn-id: http://svn.php.net/repository/php/php-src/trunk@318718 c90b9560-bf6c-de11-be94-00142212c4b1

+++ sapi/apache/mod_php5.c
@@ -529,7 +529,7 @@ static void init_request_info(TSRMLS_D)
 	SG(request_info).request_uri = r->uri;
 	SG(request_info).request_method = (char *)r->method;
 	SG(request_info).content_type = (char *) table_get(r->subprocess_env, "CONTENT_TYPE");
-	SG(request_info).content_length = (content_length ? atoi(content_length) : 0);
+	SG(request_info).content_length = (content_length ? atol(content_length) : 0);
 	SG(sapi_headers).http_response_code = r->status;
 	SG(request_info).proto_num = r->proto_num;
